build: Rewrite banner.js in TS
This commit removes the code duplication in banner.js by rewriting it in TS.
The workflow is as follows:

1. Compile banner.ts
2. rollup server/out/banner/banner.js to produce dist/server/banner.js
3. Compile the server
4. rollup server and concat banner.js to produce dist/server/index.js

In order to tree-shake unneeded code in banner.ts, a separate tsconfig
`server/banner.tsconfig.json` is needed to produce es2015 modules.
