+++ banner.rollup.config.js
@@ -1,0 +1,13 @@
+module.exports = [
+  {
+    input: 'server/out/banner/banner.js',
+    output: {
+      file: 'dist/server/banner.js',
+      format: 'cjs',
+    },
+    external: [
+      'path',
+    ],
+  },
+
+];
