+++ rollup.config.js
@@ -23,7 +23,7 @@ module.exports = [
     output: {
       file: 'dist/server/index.js',
       format: 'amd',
-      banner: fs.readFileSync('server/src/banner.js', 'utf8'),
+      banner: fs.readFileSync('dist/server/banner.js', 'utf8'),
     },
     external: [
       'fs',
