+++ server/banner.tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "target": "es2015",
+    "module": "es2015",
+    "moduleResolution": "node",
+    "sourceMap": false,
+    "strict": true,
+    "outDir": "out/banner"
+  },
+  "files": [
+    "src/banner.ts"
+  ]
+}
